Improve autocompletion in Link to CLA page
This commit improves the autocomplete experience in the "Link to CLA"
page in several ways.

In order to enable the autoSelect feature in Twitter typeahead, a custom
fork of the library has been added to the project (See
twitter/typeahead.js#1356 for more details). This feature selects
automatically the first suggestion.

A new event handler has been added on tokenfield to only allow tokens
that are present in the autosuggest engine. It is now impossible to add
random values as tokens.

A few improvements have been made with the static resources managements:

* added the webjars-locator dependency in order to have
version-agnostic links
* jquery as a webjar dependency
* use a content version strategy to have cache busting URLs and avoid
browser caching issues when pushing a new version of the app in prod

Fixes #1