Add "Loading..." indicator for repositories
Add a "Loading..." indicator and temporarily disable the search input
while the Twitter Typeahead search engine is being initialized. This
engine is initialized by a list of repositories "prefetched" from the
application. Since the API collecting this data is making several remote
calls to the Github API, it's necessary to tell the user that the search
feature is not yet available.

Note: Once prefetched, this data is kept in the browser's local storage,
until it's cleared by the application.

Fixes #5