Replace Typeahead with Select2
This commit replaces the Twitter Typeahead library, which operates on a
"input text" tag. In this case, a multiple selection "select" tag is
more appropriate.

The Select2 library dynamically loads the repositories and adds the
relevant options to the form. This option works well with pre-selected
values written by the templating engine (e.g. after a validation error).

The list of repositories is stored in the sessionStorage, which is kept
by the browser as long as the page session is opened (reloading the page
keeps the data).

Fixes #16 #6