Introduce a workaround for the CF plugin production deployment.
The CF plugin uses a deprecated API to obtain domains. Our cla.pivotal.io domain is not listed there so we need a workaround which is based on a decoration of the deployment/swap tasks.
The production route is removed before the deployment from the inactive variant (precaution, to not let cfDeploy fail) and added to the active variant once cfSwapDeployed is finished.

The CF client validates domain names in the cfDeploy and cfSwapDeployed tasks and fails if the cla.pivotal.io domain is not found. Removing the space-scoping for domain retrieval does the trick but requires a lot of auxiliary code.

Fixes gh-91.