Provide a mechanism to indicate that a change is considered obvious
Users can now add a comment (issue comment/review comment within a commit) that the change contains an "obvious fix". We no longer require the CLA to be signed. In case the CLA is signed we keep the regular flow and update the status up CLA signed. If the user is a first-time user and the PR has a comment saying "Please sign..." we add a disarming comment stating "CLA signing not necessary" and update the commit status to "CLA not required".

Fixes gh-132