Ignore the Obvious Fix words if the bot is not mentioned in comments
This commit helps with "false positives" whenever people use the words
"Obvious Fix" in a comment that was not addressed to the bot.

For example, one could ask the maintainers if the PR "can be considered
an obvious fix". This sort of comment would previously trigger the bot
to answer "this is an obvious fix", misleading the user into thinking no
CLA signature is needed.

Now such a comment will be taken into consideration if and only if the
comment starts with a mention to the bot in the `@claUserLogin ` format.

Fixes #150